adaptation du model activite/evenement et type

+++ Backend/database/migrations/2025_07_04_190555_create_avis_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->onDelete('cascade')
                 ->nullable(); 
             $table->text('contenu');
-            $table->unsignedBigInteger('etoiles')->nullable();
+            $table->decimal('etoiles',8, 1)->nullable();
             
         });
     }
